requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
faiss-cpu
numpy
python-multipart
python-dotenv
openai
pymupdf


